Add WIFI and MQTT, subscribe and apply PID parameters
Add several new files that implement wifi and mqtt
Connect to wifi and subscribe 4 topics that apply
Kp Ki Kd for valve control.
Note: needs refactoring to simplify structure

wifi.*:
    - add code for wifi support from older other project

global.cpp global.hpp:
- make servo and valveControl global for now (mqtt needs access)

config.h, credentials.h:
- mqtt and wifi config

mqtt.*:
- add custom mqtt code from older other project

mqtt.cpp:
- subscribe to 4 topics for configuring PID
- add handlers to acually update the parameters

pumpControl:
- Add missing function definitions (set config)

main:
- start wifi and mqtt

+++ fw/main/pumpControl.hpp
@@ -8,12 +8,12 @@ class ControlledValve
 public:
     ControlledValve(ServoMotor *pValve);
     void compute(float pressureDiff); // calculate and move valve to new position
-    void setKp(float Kp);
-    void setKi(float Ki);
-    void setKd(float Kd);
+    void setKp(double Kp);
+    void setKi(double Ki);
+    void setKd(double Kd);
     void reset(); // reset integral
     void getCurrentStats(uint32_t *timeLastUpdate, float *p, float *i, float *d, float *valvePos) const;
-    void getCurrentSettings(float *kp, float *ki, float *kd) const;
+    void getCurrentSettings(double *kp, double *ki, double *kd) const;
 
 
 private:
